src/retrieval/graph_retriever/requirements.txt
elasticsearch==8.17.1
sentence-transformers==3.4.1
torch==2.7.0
llama-index==0.12.36
llama-index-vector-stores-elasticsearch==0.4.3
tqdm
pytest
loguru
rapidfuzz
diskcache
jsonnet
more-itertools
pydantic
gunicorn
requests
flask
flask[async]
flask-cors
huggingface_hub==0.25.2
ftfy

